# backend/docs/swagger.json
{
  "openapi": "3.0.2",
  "info": {
      "title": "Games API TARpe22",
      "version": "1.0.0",
      "description": "# Games API Tarpe22 Hajusrakenduste näidisrakendus \n\n### Error handling\nErrors are presented using the HTTP status codes 400-599. The response include a more detailed explanation about the error in the body.\n\n```\n{\n  \"error\": \"Cleartext explanation of the error that occurred\"\n}\n```\n\n### Generic API errors\nThese error responses can be returned by any of the described endpoints.\n\n| HTTP status  |  Description |\n|---|---|\n| 400| Bad input parameters or invalid request. More information can be found in the response body.\n| 401| The consumer does not have access to the requested resource or the provided token is not valid.\n| 404| The requested resource can not be found or the user does not have access to the resource.\n| 5XX| A server error has occurred while processing the request."
  },
  "paths": {
      "/games": {
          "summary": "Games object",
          "get": {
              "tags": [
                  "GAMES"
              ],
              "responses": {
                  "200": {
                      "content": {
                          "application/json": {
                              "schema": {
                                  "type": "array",
                                  "items": {
                                      "$ref": "#/components/schemas/GameListItem"
                                  }
                              },
                              "examples": {
                                  "example": {
                                      "value": {
                                          "id": 0,
                                          "name": "Witcher 3"
                                      }
                                  }
                              }
                          }
                      },
                      "description": "Success"
                  }
              },
              "summary": "Get all game names"
          }
      },
      "/games/{id}": {
          "get": {
              "tags": [
                  "Games"
              ],
              "responses": {
                  "200": {
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Game"
                              },
                              "examples": {
                                  "example": {
                                      "value": {
                                          "id": 1,
                                          "name": "Cyberpunk 2077",
                                          "price": 39.99
                                      }
                                  },
                                  "Game without price": {
                                      "value": {
                                          "id": 3,
                                          "name": "Minecraft"
                                      }
                                  }
                              }
                          }
                      },
                      "description": "Success"
                  },
                  "400": {
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Error"
                              },
                              "examples": {
                                  "InvalidId": {
                                      "value": "{\n    error:`Id must be a number: ${req.params.id}\"`\n}\n"
                                  }
                              }
                          }
                      },
                      "description": "Invalid request\n"
                  },
                  "404": {
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Error"
                              },
                              "examples": {
                                  "IdNotfound": {
                                      "value": "{\n    error: \"Game not found\"\n}\n"
                                  }
                              }
                          }
                      },
                      "description": "Game with given id does not exist"
                  }
              },
              "summary": "Get the game with given id"
          },
          "parameters": [
              {
                  "name": "id",
                  "schema": {
                      "type": "integer"
                  },
                  "in": "path",
                  "required": true
              }
          ]
      }
  },
  "components": {
      "schemas": {
          "Game": {
              "title": "Root Type for Game",
              "description": "",
              "type": "object",
              "properties": {
                  "id": {
                      "format": "int32",
                      "type": "integer"
                  },
                  "name": {
                      "type": "string"
                  },
                  "price": {
                      "format": "double",
                      "type": "number"
                  }
              },
              "example": {
                  "id": 5,
                  "name": "Grand theft Auto V",
                  "price": 29.99
              }
          },
          "GameListItem": {
              "title": "Root Type for GameListItem",
              "description": "",
              "type": "object",
              "properties": {
                  "id": {
                      "format": "int32",
                      "type": "integer"
                  },
                  "name": {
                      "type": "string"
                  }
              },
              "example": {
                  "id": 0,
                  "name": "Minecraft"
              }
          },
          "Error": {
              "title": "Root Type for Error",
              "description": "",
              "type": "object",
              "properties": {
                  "error": {
                      "type": "string"
                  }
              },
              "example": {
                  "error": "Error message"
              }
          }
      }
  }
}